% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{NCT_convert_mon_to_num}
\alias{NCT_convert_mon_to_num}
\title{Convert monetary value to numeric}
\usage{
NCT_convert_mon_to_num(vec)
}
\arguments{
\item{vec}{A string or vector of strings containing an extracted monetary value}
}
\value{
A vector of numeric values
}
\description{
Converts your monetary value string into a numeric value, cleaning and adjusting symbols.
Takes "." as a filler and changes "," into "." to represent decimals.
}
\examples{
NCT_convert_mon_to_num("R$ 1.233,50")

}
