% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{NCT_convert_perct_to_num}
\alias{NCT_convert_perct_to_num}
\title{Converts percentages to numeric}
\usage{
NCT_convert_perct_to_num(vec)
}
\arguments{
\item{vec}{A extracted string (or a vector of those) containing a percentage value}
}
\value{
A numeric value
}
\description{
Converts percentage values to numeric, also dividing them by 100 to represent percentages in a proportion.
}
\examples{
NCT_convert_perct_to_num("5\%")

}
