% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{NCT_find_nold_clauses}
\alias{NCT_find_nold_clauses}
\title{Finds old and new articles}
\usage{
NCT_find_nold_clauses(df)
}
\arguments{
\item{A}{data.frame containing columns with NA, representing articles that were added or taken out}
}
\value{
A named character vector with these articles conditions
}
\description{
Finds old and new articles from a data.frame containing ordered articles of these NCT.
Recommended to be used with `dplyr` `bind_rows` to create your data.frames
}
