% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{NCT_generate_file_name}
\alias{NCT_generate_file_name}
\title{Generate file name for your data.frame}
\usage{
NCT_generate_file_name(sind, negot, df, row_chos)
}
\arguments{
\item{sind}{A union name. Our convention is to use the labor unions name here, if there's many of them, select the most important one.}

\item{negot}{Another part name. It can be an enterprise, federation, etc.}

\item{df}{Your data.frame to be exported. Be careful so that it follows our standards. To see it, check our github page.}

\item{row_chos}{A row to isolate and store as a data.frame.}
}
\description{
Generates a file name for your data.frame, taking a year from each of its row names.
It's intended to be an easy way to automate and standardize our files, making it easy to find them in our github.
}
\examples{
for (i in 1:nrow(CCT_df)){
write.csv(x = CCT_df[i,],
          file = NCT_generate_file_name("STIAC", "SINDITRIGO", CCT_df, i),
          fileEncoding = "utf8")
}

}
