% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{NCT_generate_ggplot2_df}
\alias{NCT_generate_ggplot2_df}
\title{Generate ggplot data.frame}
\usage{
NCT_generate_ggplot2_df(obj_name, vec)
}
\arguments{
\item{obj_name}{A string to represent your object analyzed}

\item{vec}{A named vector containing its time period as names and numeric values}
}
\value{
A ggplot2 like data.frame to bind and create time series plots
}
\description{
Create a new data.frame from a named vector containing some periodically variation.
Uses these vectors names as their period (generally, if we are talking about NCTs, a year)
}
\examples{
piso_salarial <- c("2012" = 2343, "2013" = 2566, "2014" = 2777)
NCT_generate_ggplot2_df("Piso", piso_salarial)

}
