% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{NCT_get_clauses}
\alias{NCT_get_clauses}
\title{Get a NCT clauses}
\usage{
NCT_get_clauses(vec, n, pos)
}
\arguments{
\item{vec}{A NCT character vector, usually from a NCT_get_NCT object}

\item{n}{The order of that clause in your NCT}

\item{pos}{A blueprint of where all clauses start in your vector}
}
\value{
A named character vector containing an article from your NCT
}
\description{
Filter a clause/article from a NCT and names it
}
