% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{NCT_get_complete_NCT}
\alias{NCT_get_complete_NCT}
\title{Get an organized NCT}
\usage{
NCT_get_complete_NCT(path)
}
\arguments{
\item{Either}{a Mediador link or a file path to a NCT}
}
\value{
An organized named character vector
}
\description{
Wrapper of our original NCT scraping functions, working as an end-to-end function
}
