% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{NCT_get_monetary_values}
\alias{NCT_get_monetary_values}
\title{Extract monetary values}
\usage{
NCT_get_monetary_values(vec)
}
\arguments{
\item{vec}{A string or vector of strings with monetary values (R$)}
}
\value{
A list of strings containing your monetary value(s)
}
\description{
Extracts monetary values (in R$) from a string or a vector of strings. Recommended to study wages.
}
\examples{
NCT_get_monetary_values(CCT_df$`SALARIO NORMATIVO`)
NCT_get_monetary_values(CCT[`SALARIO NORMATIVO`])

}
