% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{NCT_get_NCT}
\alias{NCT_get_NCT}
\title{Get a NCT text}
\usage{
NCT_get_NCT(path)
}
\arguments{
\item{path}{Either a Mediador link or a file path to a NCT}
}
\value{
A character vector
}
\description{
Download raw text from a Collective Bargaining Contract (Negociacao Coletiva de Trabalho - NCT) with some formatting
}
\examples{
NCT_get_NCT("http://www3.mte.gov.br/sistemas/mediador/Resumo/ResumoVisualizar?NrSolicitacao=MR039775/2018")
NCT_get_NCT("SENGE_SICEPOT_2018.html")

}
