% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{NCT_get_percent_value}
\alias{NCT_get_percent_value}
\title{Extracts percentages values}
\usage{
NCT_get_percent_value(vec)
}
\arguments{
\item{vec}{A string or vector of strings containing percentages values}
}
\value{
A list of strings containing percentage(s) value(s)
}
\description{
Extracts percentages values from a character string. Recommended to study wages correction.
}
\examples{
NCT_get_percent_value(CCT_df$`REAJUSTE SALARIAL`)

}
