% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{NCT_organize_NCT}
\alias{NCT_organize_NCT}
\title{Organize NCT}
\usage{
NCT_organize_NCT(NCT)
}
\arguments{
\item{NCT}{A character vector containing all your articles and extracted from NCT_get_NCT}
}
\description{
Filter all articles and paste them together in a named character vector
}
